more tools to help debug kron with parameters
